+++ vit-servicing-station-tests/src/common/mod.rs
@@ -1,5 +1,6 @@
 pub mod clients;
 pub mod data;
+pub mod db;
 pub mod paths;
 pub mod server;
 pub mod startup;
