+++ vit-servicing-station-tests/src/common/startup/db.rs
@@ -5,14 +5,14 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::str::FromStr;
 use thiserror::Error;
-use vit_servicing_station_lib::db::{
-    models::{api_tokens::APITokenData, proposals::Proposal},
-    schema::{api_tokens, proposals, voteplans},
-};
+use vit_servicing_station_lib::db::models::{api_tokens::APITokenData, proposals::Proposal};
 
 use std::fs;
 
-use crate::common::paths::MIGRATION_DIR;
+use crate::common::{
+    db::{DbInserter, DbInserterError},
+    paths::MIGRATION_DIR,
+};
 
 pub struct DbBuilder {
     migrations_folder: Option<PathBuf>,
@@ -22,13 +22,8 @@ pub struct DbBuilder {
 
 impl DbBuilder {
     pub fn new() -> Self {
-        let migration_dir = PathBuf::from_str(MIGRATION_DIR).unwrap();
-
-        println!("EXE: {:?}", std::env::current_exe());
-        println!("{:?}", fs::canonicalize(&migration_dir));
-
         Self {
-            migrations_folder: Some(migration_dir),
+            migrations_folder: Some(PathBuf::from_str(MIGRATION_DIR).unwrap()),
             token: None,
             proposals: None,
         }
@@ -83,61 +78,14 @@ impl DbBuilder {
 
     fn try_insert_token(&self, connection: &SqliteConnection) -> Result<(), DbBuilderError> {
         if let Some(token) = &self.token {
-            let values = (
-                api_tokens::dsl::token.eq(&(*token.token.as_ref())),
-                api_tokens::dsl::creation_time.eq(token.creation_time),
-                api_tokens::dsl::expire_time.eq(token.expire_time),
-            );
-
-            diesel::insert_into(api_tokens::table)
-                .values(values)
-                .execute(connection)
-                .map_err(DbBuilderError::DieselError)?;
+            DbInserter::new(connection).insert_token(token)?;
         }
         Ok(())
     }
 
     fn try_insert_proposals(&self, connection: &SqliteConnection) -> Result<(), DbBuilderError> {
         if let Some(proposals) = &self.proposals {
-            for proposal in proposals {
-                let values = (
-                    proposals::proposal_id.eq(proposal.proposal_id.clone()),
-                    proposals::proposal_category
-                        .eq(proposal.proposal_category.category_name.clone()),
-                    proposals::proposal_title.eq(proposal.proposal_title.clone()),
-                    proposals::proposal_summary.eq(proposal.proposal_summary.clone()),
-                    proposals::proposal_problem.eq(proposal.proposal_problem.clone()),
-                    proposals::proposal_solution.eq(proposal.proposal_solution.clone()),
-                    proposals::proposal_public_key.eq(proposal.proposal_public_key.clone()),
-                    proposals::proposal_funds.eq(proposal.proposal_funds),
-                    proposals::proposal_url.eq(proposal.proposal_url.clone()),
-                    proposals::proposal_files_url.eq(proposal.proposal_files_url.clone()),
-                    proposals::proposer_name.eq(proposal.proposer.proposer_name.clone()),
-                    proposals::proposer_contact.eq(proposal.proposer.proposer_email.clone()),
-                    proposals::proposer_url.eq(proposal.proposer.proposer_url.clone()),
-                    proposals::chain_proposal_id.eq(proposal.chain_proposal_id.clone()),
-                    proposals::chain_proposal_index.eq(proposal.chain_proposal_index),
-                    proposals::chain_vote_options.eq(proposal.chain_vote_options.as_csv_string()),
-                    proposals::chain_voteplan_id.eq(proposal.chain_voteplan_id.clone()),
-                );
-                diesel::insert_into(proposals::table)
-                    .values(values)
-                    .execute(connection)?;
-
-                // insert the related fund voteplan information
-                let voteplan_values = (
-                    voteplans::chain_voteplan_id.eq(proposal.chain_voteplan_id.clone()),
-                    voteplans::chain_vote_start_time.eq(proposal.chain_vote_start_time),
-                    voteplans::chain_vote_end_time.eq(proposal.chain_vote_end_time),
-                    voteplans::chain_committee_end_time.eq(proposal.chain_committee_end_time),
-                    voteplans::chain_voteplan_payload.eq(proposal.chain_voteplan_payload.clone()),
-                    voteplans::fund_id.eq(proposal.fund_id),
-                );
-
-                diesel::insert_into(voteplans::table)
-                    .values(voteplan_values)
-                    .execute(connection)?;
-            }
+            DbInserter::new(connection).insert_proposals(proposals)?;
         }
         Ok(())
     }
@@ -168,8 +116,8 @@ impl Default for DbBuilder {
 
 #[derive(Error, Debug)]
 pub enum DbBuilderError {
-    #[error("internal diesel error")]
-    DieselError(#[from] diesel::result::Error),
+    #[error("cannot insert data")]
+    DbInserterError(#[from] DbInserterError),
     #[error("Cannot open or create database")]
     CannotCreateDatabase(#[from] rusqlite::Error),
     #[error("Cannot initialize on temp directory")]
